# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Server Sorter",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/servers": {
      "get": {
        "summary": "Servers",
        "operationId": "servers_api_v1_servers_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500,
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ranking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rankings": {
      "post": {
        "summary": "Rankings",
        "operationId": "rankings_api_v1_rankings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ranking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons": {
      "post": {
        "summary": "Comparisons",
        "operationId": "comparisons_api_v1_comparisons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServerRecord"
                  },
                  "type": "array",
                  "title": "Response Comparisons Api V1 Comparisons Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations": {
      "get": {
        "summary": "Evaluation History",
        "operationId": "evaluation_history_api_v1_evaluations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationReport"
                  },
                  "type": "array",
                  "title": "Response Evaluation History Api V1 Evaluations Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Evaluations",
        "operationId": "evaluations_api_v1_evaluations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EvaluationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{identifier}": {
      "get": {
        "summary": "Evaluation Report",
        "operationId": "evaluation_report_api_v1_evaluations__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections": {
      "get": {
        "summary": "Collections",
        "operationId": "collections_api_v1_collections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Selection"
                  },
                  "type": "array",
                  "title": "Response Collections Api V1 Collections Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Collection",
        "operationId": "create_collection_api_v1_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Selection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{identifier}/export": {
      "get": {
        "summary": "Export Collection",
        "operationId": "export_collection_api_v1_collections__identifier__export_get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionExport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/import": {
      "post": {
        "summary": "Import Collection",
        "operationId": "import_collection_api_v1_collections_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionExport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Selection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/configurations": {
      "post": {
        "summary": "Configurations",
        "operationId": "configurations_api_v1_configurations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Configuration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Configurations Api V1 Configurations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/versions/activate": {
      "post": {
        "summary": "Activate Versions",
        "operationId": "activate_versions_api_v1_versions_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Activate Versions Api V1 Versions Activate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backups/{identifier}": {
      "get": {
        "summary": "Verify Backup",
        "operationId": "verify_backup_api_v1_backups__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Verify Backup Api V1 Backups  Identifier  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/versions/diff": {
      "get": {
        "summary": "Version Diff",
        "operationId": "version_diff_api_v1_versions_diff_get",
        "parameters": [
          {
            "name": "before",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Before"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "title": "Response Version Diff Api V1 Versions Diff Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CollectionRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Ids"
          },
          "snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          }
        },
        "type": "object",
        "required": ["name", "ids"],
        "title": "CollectionRequest"
      },
      "CompareRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 2,
            "title": "Ids"
          },
          "snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          }
        },
        "type": "object",
        "required": ["ids"],
        "title": "CompareRequest"
      },
...
      "EvaluationReport": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "dataset_sha256": {
            "type": "string",
            "title": "Dataset Sha256"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "execution_mode": {
            "type": "string",
            "title": "Execution Mode",
            "default": "fixture"
          },
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "baseline"
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseResult"
            },
            "type": "array",
            "title": "Cases"
          },
          "summary": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Summary"
          },
          "slices": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "type": "object",
            "title": "Slices"
          },
          "notice": {
            "type": "string",
            "title": "Notice",
            "default": "Fixture regression results. Live model quality is unmeasured."
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_version",
          "dataset_sha256",
          "source_sha256",
          "provenance",
          "cases",
          "summary"
        ],
        "title": "EvaluationReport"
      },
      "EvaluationRequest": {
        "properties": {
          "profile": {
            "type": "string",
            "enum": [
              "baseline",
              "demo-balanced",
              "demo-fast",
              "demo-timeout",
              "demo-malformed",
              "demo-rate-limited",
              "demo-unavailable",
              "openrouter",
              "litellm"
            ],
            "title": "Profile",
            "default": "baseline"
          }
        },
        "type": "object",
        "title": "EvaluationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RankRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "title": "Query",
            "default": ""
          },
          "filters": {
            "$ref": "#/components/schemas/Filters"
          },
          "snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          },
          "configuration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          },
          "profile": {
            "type": "string",
            "enum": [
              "baseline",
              "demo-balanced",
              "demo-fast",
              "demo-timeout",
              "demo-malformed",
              "demo-rate-limited",
              "demo-unavailable",
              "openrouter",
              "litellm"
            ],
            "title": "Profile",
            "default": "baseline"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RankRequest"
      },
...
      "Ranking": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "application_version": {
            "type": "string",
            "title": "Application Version",
            "default": "0.1.0"
          },
          "snapshot": {
            "type": "string",
            "title": "Snapshot"
          },
          "configuration": {
            "type": "string",
            "title": "Configuration",
            "default": ""
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "evidence-order-v1"
          },
          "model_profiles_version": {
            "type": "string",
            "title": "Model Profiles Version",
            "default": "profiles-v1"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version",
            "default": "bm25-evidence-v1"
          },
          "sqlite_version": {
            "type": "string",
            "title": "Sqlite Version",
            "default": "3.50.4"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "deterministic"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "filters": {
            "$ref": "#/components/schemas/Filters"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RankedServer"
            },
            "type": "array",
            "title": "Results"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "model_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Model Metadata"
          }
        },
        "type": "object",
        "required": ["snapshot", "as_of", "query", "filters", "results"],
        "title": "Ranking"
      },
...
      "ServerRecord": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "pattern": "^[a-zA-Z0-9._/-]+$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "version": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Version"
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description"
          },
          "category": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Category"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Tags"
          },
          "transport": {
            "type": "string",
            "enum": ["stdio", "streamable-http", "unknown"],
            "title": "Transport",
            "default": "unknown"
          },
          "auth": {
            "type": "string",
            "enum": ["none", "api_key", "oauth", "unknown"],
            "title": "Auth",
            "default": "unknown"
          },
          "deployment": {
            "type": "string",
            "enum": ["local", "remote", "unknown"],
            "title": "Deployment",
            "default": "unknown"
          },
          "license": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "status": {
            "type": "string",
            "enum": ["active", "deprecated", "unknown"],
            "title": "Status",
            "default": "unknown"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Evidence"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "description",
          "category",
          "tags",
          "updated_at",
          "evidence"
        ],
        "title": "ServerRecord"
      },
...
    }
  }
}